Extracted common mock objects from tests.
