+++ nh_eobs_mental_health/tests/report_nh_clinical_observation_report/test_get_first_refusal_column_data.py
@@ -1,7 +1,10 @@
 # -*- coding: utf-8 -*-
+from copy import deepcopy
 from datetime import datetime
 
 from openerp.tests.common import TransactionCase
+
+from . import patient_refusal_event_mocks
 
 
 class TestGetFirstRefusalColumnData(TransactionCase):
@@ -12,44 +15,8 @@ class TestGetFirstRefusalColumnData(TransactionCase):
         super(TestGetFirstRefusalColumnData, self).setUp()
         self.report_model = self.env['report.nh.clinical.observation_report']
 
-        self.mock_refusal_episode = {
-            'count': 1,
-            'first_refusal_date_terminated': '2017-01-03 17:49:11.36621',
-            'freq_date_terminated': None,
-            'freq_state': None,
-            'freq_terminate_uid': None,
-            'review_date_terminated': None,
-            'review_state': None,
-            'review_terminate_uid': None,
-            'spell_activity_id': 10
-        }
-        # mock_refusal_episode_second = {
-        #     'count': 1,
-        #     'first_refusal_date_terminated': '2017-01-03 17:49:1#3.36621',
-        #     'freq_date_terminated': None,
-        #     'freq_state': None,
-        #     'freq_terminate_uid': None,
-        #     'review_date_termninated': None,
-        #     'review_state': None,
-        #     'review_terminate_uid': None,
-        #     'spell_activity_id': 10
-        # }
-        # mock_refusal_episode_third = {
-        #     'count': 1,
-        #     'first_refusal_date_terminated': '2017-01-03 17:49:14.36621',
-        #     'freq_date_terminated': None,
-        #     'freq_state': None,
-        #     'freq_terminate_uid': None,
-        #     'review_date_termninated': None,
-        #     'review_state': None,
-        #     'review_terminate_uid': None,
-        #     'spell_activity_id': 10
-        # }
-        # self.mock_refusal_episodes = [
-        #     mock_refusal_episode_third,
-        #     mock_refusal_episode_first,
-        #     mock_refusal_episode_second
-        # ]
+        self.mock_refusal_episode = \
+            deepcopy(patient_refusal_event_mocks.mock_refusal_episode_first)
 
     def call_test(self, arg=False):
         if arg is False:
